requirements.txt
django==3.1
django-cors-headers==3.5.0
djangorestframework==3.12.2
mysqlclient
djangorestframework-simplejwt==4.7.2